refactor: ProjectModalをシンプル化し、OngoingIndicatorを分離
